Attempt to tile the currently's focused window
Adding functionality to attempt tiling from the window that is "focused"
according to Windows. Fall back to tiling on root if the last focused
window cannot be found. Unfortuantely, the fall back will happen quite
frequently under the current implementation because there are _a lot_ of
different things that can trigger a focus change in Windows, and there
is no current filtering code on the focus change callback code to get
rid of them. A step in the right direction though, probably.